// yanay-tevet-web/src/generated-files/api/shopping-lists/types.gen.ts
export type ShoppingListSchema = {
    /**
     * Id
     */
    id: number;
    /**
     * Name
     */
    name: string;
    /**
     * Owner Id
     */
    owner_id: number;
    /**
     * Owner Username
     */
    owner_username: string;
    /**
     * Member Count
     */
    member_count: number;
    /**
     * Unchecked Count
     */
    unchecked_count: number;
    /**
     * Checked Count
     */
    checked_count: number;
    /**
     * Preview Items
     */
    preview_items: Array<string>;
    /**
     * Created At
     */
    created_at: string;
    /**
     * Updated At
     */
    updated_at: string;
};
// ...
export type ShoppingItemSchema = {
    /**
     * Id
     */
    id: number;
    /**
     * Shopping List Id
     */
    shopping_list_id: number;
    /**
     * Name
     */
    name: string;
    /**
     * Is Checked
     */
    is_checked: boolean;
    /**
     * Checked At
     */
    checked_at: string | null;
    /**
     * Order
     */
    order: number;
    /**
     * Created At
     */
    created_at: string;
    /**
     * Updated At
     */
    updated_at: string;
};